Moved DetailData component files to a folder
To match organization of other components
