+++ src/App.jsx
@@ -1,6 +1,6 @@
 import SummaryMetric from './components/SummaryMetric/SummaryMetric';
 import Header from './components/Header/Header';
-import DetailData from './DetailData';
+import DetailData from './components/DetailData/DetailData';
 
 function App() {
   return (
